pyproject.toml
[project]
name = "llm-usage"
version = "0.1.0"
description = "track LLM CLI usage windows with ambient keyboard + voice feedback"
authors = [
    {name = "Peter Gomez", email = "fadedlamp42@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "PyYAML>=6.0",
    "prism-shared-python",
]

[tool.poetry.dependencies]
prism-shared-python = {path = "../../../prism/repositories/prism-shared-python", develop = true}

[tool.poetry.group.dev.dependencies]
ruff = ">=0.15"
pre-commit = ">=4.0"

[tool.poetry.scripts]
llm-usage = "llm_usage.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# -- ruff (linting + formatting) --

[tool.ruff]
target-version = "py313"
line-length = 99

[tool.ruff.lint]
select = [
    "F",      # pyflakes - unused imports, undefined names
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "I",      # isort - import ordering
    "N",      # pep8-naming
    "UP",     # pyupgrade - modernize syntax
    "B",      # flake8-bugbear - common pitfalls
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
    "C4",     # flake8-comprehensions
    "PIE",    # flake8-pie - misc lints
    "PT",     # flake8-pytest-style (if tests ever exist)
    "RET",    # flake8-return
    "TCH",    # flake8-type-checking - move imports behind TYPE_CHECKING
    "LOG",    # flake8-logging
    "T201",   # print() statements - use structlog instead
]
ignore = [
    "E501",   # line length handled by formatter
    "UP031",  # we prefer %(name)s style formatting per project convention
    "N806",   # ObjC class lookups produce uppercase locals, that's fine
    "RET504", # unnecessary assignment before return (readability preference)
    "SIM108", # ternary instead of if/else (readability preference)
]

[tool.ruff.lint.isort]
known-first-party = ["llm_usage"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
